code cleanup/refactoring (focused on testing/playground related code)
